before change  wallet table

+++ Application/Interfaces/IWalletService.cs
@@ -33,10 +33,10 @@ namespace Application.Services.Interfaces
         Task<WalletViewModel> GetWalletById(ulong id);
 
         //Create New Wallet Transaction For Redirext To The Bank Portal
-        Task CreateNewWalletTransactionForRedirextToTheBankPortal(int userId, int price, GatewayType gateway, string authority, string description, ulong? requestId);
+        Task CreateNewWalletTransactionForRedirextToTheBankPortal(int userId, int price, GatewayType gateway, string authority, string description, int? requestId);
 
         //Find Wallet Transaction For Redirect To The Bank Portal 
-        Task<Wallet?> FindWalletTransactionForRedirectToTheBankPortal(int userId, GatewayType gateway, ulong? requestId, string authority, int amount);
+        Task<Wallet?> FindWalletTransactionForRedirectToTheBankPortal(int userId, GatewayType gateway, int? requestId, string authority, int amount);
 
         //Update Wallet And Calculate User Balance After Banking Payment
         Task UpdateWalletAndCalculateUserBalanceAfterBankingPayment(Wallet wallet);
