+++ Data/Repository/WalletRepository.cs
@@ -163,7 +163,7 @@ namespace DoctorFAM.Data.Repository
         }
 
         //Find Wallet Transaction For Redirect To The Bank Portal 
-        public async Task<Wallet?> FindWalletTransactionForRedirectToTheBankPortal(int userId, GatewayType gateway, ulong? requestId, string authority, int amount)
+        public async Task<Wallet?> FindWalletTransactionForRedirectToTheBankPortal(int userId, GatewayType gateway, int? requestId, string authority, int amount)
         {
             return await _context.Wallets.Include(p => p.WalletData).FirstOrDefaultAsync(p => !p.IsDelete && !p.IsFinally && p.UserId == userId && p.GatewayType == gateway
                                                                                                 && p.WalletData.TrackingCode == authority &&
